fix(doom/f_wipe): compute wipe_initMelt copy size in size_t
wipe_initMelt does memcpy(wipe_scr, wipe_scr_start, width*height). width and
height are int, so the byte count is computed in 32-bit and only widened to
memcpy's size_t afterwards — it can wrap before the conversion. CodeQL flags it
as cpp/integer-multiplication-cast-to-long (CWE-190).

Cast the first operand to size_t so the product is 64-bit. No behavioural change
at real screen dimensions.

+++ kernel/doom_src/f_wipe.c
@@ -138,7 +138,7 @@ wipe_initMelt
     int i, r;
     
     // copy start screen to main screen
-    memcpy(wipe_scr, wipe_scr_start, width*height);
+    memcpy(wipe_scr, wipe_scr_start, (size_t)width*height);   // 64-bit multiply before widening to size_t
     
     // makes this wipe faster (in theory)
     // to have stuff in column-major format
